echo_app_python/requirements.txt
requests
google-auth
Flask
gunicorn